feat(auth): singleton AuthProvider — one onAuthStateChange, locationId/orgId in context
- Moves all auth logic (buildUser, onAuthStateChange, state) to AuthProvider
  mounted once at root layout, persisting across client-side navigations
- useAuth() is now useContext(AuthContext) — same API, zero redundant subscriptions
- locationId + orgId exposed on context with single DEV fallback (removes 4 per-page copies)
- Removes TOKEN_REFRESHED subscription from useDashboardData (kills duplicate RPC bursts)
- Adds Sentry.captureMessage safety net when activeMembership would drop to null
