+++ app/dashboard/reconcile/page.tsx
@@ -135,14 +135,10 @@ function DataTable({
 
 export default function ReconcilePage() {
   const router         = useRouter()
-  const { user, isLoading, error: authError } = useAuth()
+  const { user, isLoading, error: authError, locationId, orgId } = useAuth()
 
-  const DEV_FALLBACK_LOCATION = 'bbbbbbbb-0000-0000-0000-000000000001'
-  const DEV_FALLBACK_ORG      = 'aaaaaaaa-0000-0000-0000-000000000001'
-  const isDev     = process.env.NODE_ENV === 'development'
-  const locationId = user?.activeMembership?.location_id ?? (isDev ? DEV_FALLBACK_LOCATION : null)
-  const orgId      = user?.activeMembership?.org_id      ?? (isDev ? DEV_FALLBACK_ORG : null)
-  const orgName    = user?.activeMembership?.organization?.name ?? 'Dashboard'
+  const isDev   = process.env.NODE_ENV === 'development'
+  const orgName = user?.activeMembership?.organization?.name ?? 'Dashboard'
 
   const [from,   setFrom]   = useState(firstOfMonth())
   const [to,     setTo]     = useState(today())
