+++ app/providers.tsx
@@ -1,0 +1,7 @@
+'use client'
+
+import { AuthProvider } from '@/providers/AuthProvider'
+
+export function Providers({ children }: { children: React.ReactNode }) {
+  return <AuthProvider>{children}</AuthProvider>
+}
